refactor(webex): expose WebexListener on the user platform
Move the Mercury WebSocket listener core (listener, id-normalizer, wdm-discovery) from the webexbot platform into the webex user platform and rename it WebexListener, matching the cross-platform naming convention (WebexListener for users, WebexBotListener for bots — like SlackListener/SlackBotListener).

WebexBotListener is kept as a backward-compatible alias of WebexListener, so the webexbot public API and tests are unchanged. WebexClient gains a public getToken() so new WebexListener(client) works. Dependency direction is preserved (webexbot -> webex).
