+++ src/platforms/webex/index.ts
@@ -1,5 +1,9 @@
 export { WebexClient } from './client'
 export { WebexCredentialManager } from './credential-manager'
+export { fromRestId, toRestId } from './id-normalizer'
+export type { WebexRestIdType } from './id-normalizer'
+export { WebexListener } from './listener'
+export type { WebexListenerClient, WebexListenerEventMap, WebexListenerOptions } from './listener'
 export { WebexTokenExtractor } from './token-extractor'
 export type { ExtractedWebexToken } from './token-extractor'
 export { WebexError } from './types'
