+++ src/platforms/webexbot/types.ts
@@ -1,12 +1,6 @@
-import type {
-  AttachmentAction,
-  DecryptedMessage,
-  DeletedMessage,
-  HandlerStatus,
-  MembershipActivity,
-  RoomActivity,
-} from 'webex-message-handler'
 import { z } from 'zod'
+
+export type { WebexListenerEventMap as WebexBotListenerEventMap } from '../webex/listener'
 
 export interface WebexBotEntry {
   bot_id: string
@@ -55,27 +49,3 @@ export const WebexBotCredentialsSchema = z.object({
   bot_id: z.string(),
   bot_name: z.string(),
 })
-
-/**
- * ID contract: every REST-resource ID on these events (person, room, message,
- * and attachment-action IDs) is emitted in Webex REST ID format, directly
- * comparable with IDs returned by `WebexBotClient`. Mercury-only activity IDs
- * (`MembershipActivity.id`, `RoomActivity.id`) stay raw, and the original
- * Mercury payload is preserved under `.raw` (except `DeletedMessage`, whose
- * upstream Mercury event omits the full activity). Decode any REST ID with
- * {@link fromRestId}.
- */
-export interface WebexBotListenerEventMap {
-  message_created: [event: DecryptedMessage]
-  message_updated: [event: DecryptedMessage]
-  message_deleted: [event: DeletedMessage]
-  membership_created: [event: MembershipActivity]
-  attachment_action: [event: AttachmentAction]
-  room_created: [event: RoomActivity]
-  room_updated: [event: RoomActivity]
-  webex_event: [event: DecryptedMessage | DeletedMessage | MembershipActivity | AttachmentAction | RoomActivity]
-  connected: [info: { connected: boolean; status: HandlerStatus }]
-  reconnecting: [attempt: number]
-  disconnected: [reason: string]
-  error: [error: Error]
-}
